refactored the idt
